requirements.txt
flask
google-cloud
flask-restplus==0.9.2
requests==2.11.1


